Moved edge hierarchy command into provider pattern
