+++ azext_iot/commands.py
@@ -134,8 +134,6 @@ def load_command_table(self, _):
 
     with self.command_group("iot edge", command_type=iothub_ops) as cmd_group:
         cmd_group.command("set-modules", "iot_edge_set_modules")
-    with self.command_group("iot edge hierarchy", command_type=iothub_ops) as cmd_group:
-        cmd_group.command("create", "iot_edge_hierarchy_create")
 
     with self.command_group(
         "iot edge deployment", command_type=iothub_ops
